requirements.txt
Django~=5.1
djangorestframework~=3.15.2
django-filter~=24.3
djoser~=2.2.3